backend/requirements.txt
aiofiles>=24.1.0,<25.0
aiosqlite>=0.20.0,<1.0
easyocr>=1.7.1,<2.0
editdistance>=0.8.1,<1.0
fastapi[standard]>=0.112.2
torch>=2.4.1,<3.0
msgpack>=1.1.0,<2.0
numpy>=1.26.4,<2.0
pillow>=10.4.0,<11.0
spacy>=3.7.0,<3.8.0
pl-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pl_core_news_lg-3.7.0/pl_core_news_lg-3.7.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl
python-magic>=0.4.27,<1.0
pydantic>=2.8.2,<3.0
sentence-transformers>=3.1.0,<4.0
SQLAlchemy>=2.0.32,<3.0
tqdm>=4.66.5
transformers>=4.44.2,<5.0
watchdog>=5.0.3,<6.0
wordfreq>=3.1.1,<4.0
librosa>=0.9.2,<1.0
